remove unused constants
